requirements.txt
streamlit
openai
azure-ai-inference
PyPDF2
